% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_acc.R
\name{as_acc}
\alias{as_acc}
\title{Convert to acc}
\usage{
as_acc(x, ...)
}
\arguments{
\item{x}{A `move2` containing acceleration data as collected by EOBS or ornitella tracking devices.}

\item{...}{currently not used}
}
\description{
Convert to acc
}
